Adding magic-square solution.
Having a bit of fun, I implemented a solution that shuffles the input set
and partitions the result repeatedly until it happens upon a magic-square.
Most of the work went into re-implementing the predicates for checking
whether a given square is magic.

Very silly solution that doesn't have predictable performance
characteristics, but it seemed like the fun thing to do.

Completed in 33:50.